test(lib-versioning): split the flow executor spec into focused suites
